+++ packages/core/__tests__/stages.test.ts
@@ -26,7 +26,9 @@ function buildContext(stage: string): Partial<IContext> {
     return {
         logger: new Logger(),
         plugins: {
-            [testPlugin]: { stage }
+            [testPlugin]: [
+                { stage }
+            ]
         }
     };
 }