+++ packages/changelog/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## Recent Changes
+## `0.2.0`
+
+* Support `displayNames` option that defines display names for packages in a monorepo (e.g., "zowe-explorer-ftp-extension" -> "Zowe Explorer Extension for FTP")
+
+## `0.1.0`
 
 * License project under Apache-2.0